+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -1,0 +1,33 @@
+package lotto.domain.lotto;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import java.util.List;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class LottoTest {
+    @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
+    @Test
+    void createLottoByOverSize() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
+    @Test
+    void createLottoByDuplicatedNumber() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("생성된 로또 번호가 오름차순으로 정렬된다.")
+    @Test
+    void sortLottoNumbers() {
+        Lotto lotto = new Lotto(List.of(6, 5, 4, 3, 2, 1));
+        String actual = lotto.getSortedNumbers();
+        String expected = "[1, 2, 3, 4, 5, 6]";
+        assertThat(actual).isEqualTo(expected);
+    }
+}
