+++ src/test/java/lotto/domain/lotto/LottosTest.java
@@ -1,8 +1,7 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.domain.lotto.Lottos;
 import lotto.domain.lotto.generator.AutoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
