+++ src/test/java/lotto/domain/lotto/PurchasePriceTest.java
@@ -1,8 +1,7 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.domain.lotto.PurchasePrice;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
